Move Reader card with module
See discussion in https://github.com/Automattic/jetpack/pull/47089#discussion_r2802698274
